requirements.txt
# ===========================================
# Odoo CE Development Dependencies
# ===========================================
# These are for LOCAL development and CI/CD
# Odoo core dependencies are in the Docker image
#
# Install: pip install -r requirements.txt
# ===========================================

# ----- Pre-commit & Quality Tools -----
pre-commit==3.7.1
black==26.1.0
isort==7.0.0
flake8==7.0.0
yamllint==1.35.1

# ----- OCA Linting -----
pylint>=3.0.0,<3.1.0
pylint-odoo==9.0.5

# ----- Testing -----
pytest==8.2.2
pytest-cov==5.0.0
coverage==7.5.3

# ----- Debugging -----
debugpy==1.8.1
ipython==8.25.0

# ----- Documentation -----
sphinx==7.3.7
sphinx-rtd-theme==2.0.0

# ----- Odoo ORM Stubs (for IDE autocomplete) -----
# Uncomment if using VS Code with Pylance
# odoo-stubs==18.0.0

# ===========================================
# Cloud IDE / CI Runner Tooling
# ===========================================
# For Supabase migrations checks, edge-function
# webhook checks via HTTP, DB introspection,
# ERD export via DBML, and general ops scripting.
# NOT Odoo-runtime (Odoo runs in its own container).
# ===========================================

# ----- Core Scripting -----
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.12.0
requests==2.32.3
httpx==0.27.2
tenacity==9.0.0
rich==13.9.4
pyyaml==6.0.3
jinja2==3.1.6

# ----- Postgres / Supabase DB Ops -----
psycopg[binary]==3.2.3
sqlalchemy==2.0.46

# ----- Schema / ERD / Docs Generation -----
dbml-to-sqlalchemy==0.9.0
sqlalchemy_schemadisplay==1.3
graphviz==0.21

# ----- Schema Validation -----
jsonschema==4.23.0

# ----- CLI Ergonomics -----
typer==0.23.1
